pyproject.toml
[project]
name = "midclaw-brain"
version = "0.1.0"
description = "MidClaw Python brain — LLM, tools, simulation"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "openai>=2.21.0",
    "httpx>=0.27.0",
    "pydantic>=2.12.0",
    "python-dotenv>=1.2.0",
    "rich>=14.0.0",
    "tenacity>=9.0.0",
    "pyyaml>=6.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]
sim = []

[project.scripts]
midclaw-brain = "brain.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
